+++ app/src/main/java/com/github/droidworksstudio/launcher/ui/widgets/WidgetFragment.kt
@@ -93,7 +93,6 @@ class WidgetFragment : Fragment(),
     private fun initializeInjectedDependencies() {
         context = requireContext()
         binding.nestScrollView.hideKeyboard()
-
         binding.nestScrollView.scrollEventListener = this
     }