+++ app/src/main/java/com/github/droidworksstudio/launcher/utils/Constants.kt
@@ -8,18 +8,18 @@ import com.github.droidworksstudio.launcher.R
 object Constants {
     const val PACKAGE_NAME = "app.easy.launcher"
     const val PACKAGE_NAME_DEBUG = "$PACKAGE_NAME.debug"
+    const val PACKAGE_PREFS = "EasyLauncher.pref"
 
     const val WIDGETS_COUNT = 2
     const val WIDGET_WEATHER = "WIDGET_WEATHER"
     const val WIDGET_BATTERY = "WIDGET_BATTERY"
 
-    const val WEATHER_UNITS = "WEATHER_UNITS"
 
     const val WEATHER_PREFS = "EasyWeather.pref"
+    const val WEATHER_UNITS = "WEATHER_UNITS"
     const val LATITUDE = "LATITUDE"
     const val LONGITUDE = "LONGITUDE"
 
-    const val PREFS_FILENAME = "EasyLauncher.pref"
     const val FIRST_LAUNCH = "FIRST_LAUNCH"
     const val SHOW_DATE = "SHOW_DATE"
     const val SHOW_TIME = "SHOW_TIME"
@@ -80,7 +80,10 @@ object Constants {
     const val LAUNCHER_FONT = "LAUNCHER_FONT"
 
     const val REQUEST_INSTALL_PERMISSION = 123
-    const val REQUEST_LOCATION_PERMISSION_CODE = 246
+    const val REQUEST_LOCATION_PERMISSION_CODE = 234
+
+    const val BACKUP_WRITE = 987
+    const val BACKUP_READ = 876
 
     const val LOCATION_DENIED = "LOCATION_DENIED"
 
@@ -175,13 +178,16 @@ object Constants {
         Bitter,
         Dotness,
         DroidSans,
-        GreatVibes,
+
+        //        GreatVibes,
         Lato,
-        Lobster,
+
+        //        Lobster,
         Merriweather,
         Montserrat,
         OpenSans,
-        Pacifico,
+
+        //        Pacifico,
         Quicksand,
         Raleway,
         Roboto,
@@ -193,13 +199,13 @@ object Constants {
                 Bitter -> ResourcesCompat.getFont(context, R.font.bitter)
                 Dotness -> ResourcesCompat.getFont(context, R.font.dotness)
                 DroidSans -> ResourcesCompat.getFont(context, R.font.open_sans)
-                GreatVibes -> ResourcesCompat.getFont(context, R.font.great_vibes)
+//                GreatVibes -> ResourcesCompat.getFont(context, R.font.great_vibes)
                 Lato -> ResourcesCompat.getFont(context, R.font.lato)
-                Lobster -> ResourcesCompat.getFont(context, R.font.lobster)
+//                Lobster -> ResourcesCompat.getFont(context, R.font.lobster)
                 Merriweather -> ResourcesCompat.getFont(context, R.font.merriweather)
                 Montserrat -> ResourcesCompat.getFont(context, R.font.montserrat)
                 OpenSans -> ResourcesCompat.getFont(context, R.font.open_sans)
-                Pacifico -> ResourcesCompat.getFont(context, R.font.pacifico)
+//                Pacifico -> ResourcesCompat.getFont(context, R.font.pacifico)
                 Quicksand -> ResourcesCompat.getFont(context, R.font.quicksand)
                 Raleway -> ResourcesCompat.getFont(context, R.font.raleway)
                 Roboto -> ResourcesCompat.getFont(context, R.font.roboto)
@@ -213,13 +219,13 @@ object Constants {
                 Bitter -> context.getString(R.string.settings_font_bitter)
                 Dotness -> context.getString(R.string.settings_font_dotness)
                 DroidSans -> context.getString(R.string.settings_font_droidsans)
-                GreatVibes -> context.getString(R.string.settings_font_greatvibes)
+//                GreatVibes -> context.getString(R.string.settings_font_greatvibes)
                 Lato -> context.getString(R.string.settings_font_lato)
-                Lobster -> context.getString(R.string.settings_font_lobster)
+//                Lobster -> context.getString(R.string.settings_font_lobster)
                 Merriweather -> context.getString(R.string.settings_font_merriweather)
                 Montserrat -> context.getString(R.string.settings_font_montserrat)
                 OpenSans -> context.getString(R.string.settings_font_opensans)
-                Pacifico -> context.getString(R.string.settings_font_pacifico)
+//                Pacifico -> context.getString(R.string.settings_font_pacifico)
                 Quicksand -> context.getString(R.string.settings_font_quicksand)
                 Raleway -> context.getString(R.string.settings_font_raleway)
                 Roboto -> context.getString(R.string.settings_font_roboto)